+++ xknx/remote_value/remote_value_control.py
@@ -4,39 +4,44 @@ Module for managing a control remote value.
 Examples are switching commands with priority control, relative dimming or blinds control commands.
 DPT 2.yyy and DPT 3.yyy
 """
-from typing import List, Optional
+from typing import TYPE_CHECKING, Any, List, Optional, Type, Union
 
-from xknx.dpt import DPTBase, DPTBinary
+from xknx.dpt import DPTArray, DPTBase, DPTBinary, DPTControlStepCode
 from xknx.exceptions import ConversionError
 
-from .remote_value import RemoteValue
+from .remote_value import AsyncCallbackType, RemoteValue
+
+if TYPE_CHECKING:
+    from xknx.telegram.address import GroupAddressableType
+    from xknx.xknx import XKNX
 
 
-class RemoteValueControl(RemoteValue):
+class RemoteValueControl(RemoteValue[DPTBinary]):
     """Abstraction for remote value used for controling."""
 
     def __init__(
         self,
-        xknx,
-        group_address=None,
-        group_address_state=None,
-        sync_state=True,
-        value_type=None,
-        device_name=None,
-        feature_name="Control",
-        after_update_cb=None,
-        invert=False,
-        passive_group_addresses: List[str] = None,
+        xknx: "XKNX",
+        group_address: Optional["GroupAddressableType"] = None,
+        group_address_state: Optional["GroupAddressableType"] = None,
+        sync_state: bool = True,
+        value_type: Optional[str] = None,
+        device_name: Optional[str] = None,
+        feature_name: str = "Control",
+        after_update_cb: Optional[AsyncCallbackType] = None,
+        passive_group_addresses: Optional[List["GroupAddressableType"]] = None,
     ):
         """Initialize control remote value."""
         # pylint: disable=too-many-arguments
-        self.invert = invert
+        if value_type is None:
+            raise ConversionError("no value type given", device_name=device_name)
+        # TODO: typing - parse from DPTControlStepCode when parse_transcoder is a classmethod
         _dpt_class = DPTBase.parse_transcoder(value_type)
-        if _dpt_class is None:
+        if _dpt_class is None or not isinstance(_dpt_class(), DPTControlStepCode):
             raise ConversionError(
                 "invalid value type", value_type=value_type, device_name=device_name
             )
-        self.dpt_class = _dpt_class
+        self.dpt_class: Type[DPTControlStepCode] = _dpt_class  # type: ignore
         super().__init__(
             xknx,
             group_address,
@@ -48,17 +53,21 @@ class RemoteValueControl(RemoteValue):
             passive_group_addresses=passive_group_addresses,
         )
 
-    def payload_valid(self, payload):
+    def payload_valid(
+        self, payload: Optional[Union[DPTArray, DPTBinary]]
+    ) -> Optional[DPTBinary]:
         """Test if telegram payload may be parsed."""
-        return isinstance(payload, DPTBinary)
+        # pylint: disable=no-self-use
+        return payload if isinstance(payload, DPTBinary) else None
 
-    def to_knx(self, value):
+    def to_knx(self, value: Any) -> DPTBinary:
         """Convert value to payload."""
-        return DPTBinary(self.dpt_class.to_knx(value, invert=self.invert))
+        return DPTBinary(self.dpt_class.to_knx(value))
 
-    def from_knx(self, payload):
+    def from_knx(self, payload: DPTBinary) -> Any:
         """Convert current payload to value."""
-        return self.dpt_class.from_knx(payload.value, invert=self.invert)
+        # TODO: DPTBinary.value is int - DPTBase.from_knx requires Tuple[int, ...] - maybe use bytes
+        return self.dpt_class.from_knx((payload.value,))
 
     @property
     def unit_of_measurement(self) -> Optional[str]:
@@ -68,4 +77,4 @@ class RemoteValueControl(RemoteValue):
     @property
     def ha_device_class(self) -> Optional[str]:
         """Return a string representing the home assistant device class."""
-        return getattr(self.dpt_class, "ha_device_class", None)
+        return getattr(self.dpt_class, "ha_device_class", None)  # type: ignore